feat(mobile): gate batch launch behind a pre-brew preparation screen
Realises the launch gate (UC6 of the brew-prep conception #1248). Terrain
testing showed the standalone "Vérifier mes ingrédients" button on the recipe
Overview tab dead-ended with no onward action, so the readiness check now
lives on the path to launching.

- Rename the recipe sticky CTA "Lancer un brassin" -> "Préparer mon brassin";
  it opens a pre-launch preparation screen instead of starting a batch directly.
- BrewPrepScreen (replaces IngredientReadinessScreen, route
  /recipes/[id]/prepare) hosts the ingredient checklist and a gated
  "Lancer le brassage" CTA: the irreversible batch start is disabled until the
  checklist is complete and guarded by a confirmation dialog. The startBatch
  mutation moves here.
- Remove the Overview "Vérifier mes ingrédients" entry point.
- Gate covers the ingredient checklist only this slice; the equipment checklist
  (A3) will fold into the same screen and extend the gate to both checklists.
